exec: add Dyno type
Add the Dyno type for managing the lifecycle of a dyno container process
group. Add the standard service interface for dyno process control. Run
the subprocess as a process group.

Setup the environment for the dyno process group via the Env field. If
Env is nil, use the local process' environment instead.

Add the CleanedEnv function for dropping the environment variables used
for internal configuration via envdecode.

Forward process lifecycle related signals to the dyno process group
instead of handling them in the current process. Add graceful shutdown
where by a call to Stop first tries to stop the dyno process group via a
SIGTERM. If the dyno processes do not exit before ShutdownPeriod has
elapsed, the process is then killed with SIGKILL.

Mark the current process as the init process for all child processes,
which orphaned child processes to be re-parented to it. Wire up SIGCHLD
signals and reap zombied child processes when they have exited. Avoid
reaping the main child process because the os/exec.Cmd.Wait() call will
reap it.

When Dyno.LoadSeccomp is set, load a seccomp profile that blocks
unshare(2) syscalls.

Change the child process capabilities if Dyno.Capabilities is set. Only
the capabilities in the slice are set. All sets are modified to be the
same: Permitted, Inheritable, Effective, Bounding, and Ambient.

Switch the uid/gid (real, effective, and saved) for the child processes.

Add a field to toggle on the hidepid=2 mount flag for the /proc
filesystem. This prevents the dyno process from seeing process from
other users in `ps` output. With this flag enabled the dyno cannot see
it's init process.